utils were transferred to separated folder. update dependent code after it
