+++ benchmark-hugging-face/distilbert_base_cased_distilled_squad_test.py
@@ -3,7 +3,7 @@ from functools import partial
 
 import onnx
 
-from utils import perf_test, get_distilbert_inputs
+from utils.utils import perf_test, get_distilbert_inputs
 
 if __name__ == "__main__":
   class MyFormatter(argparse.ArgumentDefaultsHelpFormatter, argparse.RawDescriptionHelpFormatter):
@@ -46,9 +46,9 @@ if __name__ == "__main__":
   benchmark_test = partial(perf_test, iters_number = args.iters_number, model_name = "Distilbert-base-cased distilled by SQuAD")
 
   if(args.tvm):
-    from tvm_utils import tvm_test
+    from utils.tvm_utils import tvm_test
     tvm_test(benchmark_test, onnx_model, encoded_inputs, opt_level, args.target, args.target, freeze)
 
   if args.ort:
-    from ort_utils import ort_test
+    from utils.ort_utils import ort_test
     ort_test(benchmark_test, args.model_path, encoded_inputs)
